Remove TxNamespaceHandler.getTransactionManager()
Checked the tx: schemas back to 2.0, and every one of them has had a
default=transactionManager on both the tx:annotation-driven and
tx:advice elements' transaction-manager attributes.

This was probably something added early on and never removed.

I'm already specifying this default for the no-arg constructor on
TxAnnotationDriven, which already means duplication with the default
declared in the XSD.  I'd rather it not be duplicated in *three* places, thus the deletion.

Note that if, somehow, the call to element.getAttribute(transaction-manager)
were to come back null, TxAnnotationDriven allows for a null parameter in its
String constructor and still falls back to the default name.  So we shouldn't
see a regression in any case.