Introduce Default[Formatting]ConversionService
Intended for convenient programmatic access to a conversion services
configured with default converters and formatters.

Prior to this change, the [Formatting]ConversionServiceFactoryBean
and ConversionServiceFactory classes contained much of the default
registration logic.

These new Default* implementations expose static 'addDefaultConverters()'
and 'addDefaultFormatters()' methods respectively, allowing for ad-hoc
use by any GenericConversionService or FormattingConversionService
implementation.

This change was driven by the need to avoid use of the associated
FactoryBean types within @Bean methods when configuring the container
programatically in @Configuration classes.

Going forward, FactoryBeans should be considered a necessary convenience
for XML configuration, but emphasis should be placed on ease of use
within @Bean methods.  Use of FactoryBean types within @Bean methods
is problematic and inelegant and to be discouraged; this requires of
course that we provide proper alternatives.