Add tests to reproduce SPR-7979
spr7979_bootstrappedFromACAC passes; spr7979_bootstrappedFromGXAC fails.

The reason is that the user-specified type 'Foo' cannot be classloaded
when operating from an XML context, while it can from a code-based
context. The root cause is not yet clear, though it strongly appears to
be a mismatch of the ClassLoader used in the different situation.

See comment in ComponentScanAnnotationParser which instructs how to
debug and replicate this error when trying to parse the 'includeFilters'
attribute of the @ComponentScan annotation.

Note that this is a pretty specific situation: (1) an annotation having
attributes that nest annotations (2) that specify user types as
attribute values (3) being parsed from an XML application context.

Issue: SPR-7979