Refactor @EnableAsync to be ImportSelector-based
@EnableAsync's #mode attribute determines whether async execution
advice is to be applied through proxying or AspectJ weaving.

This means registering AsyncAnnotationBeanPostProcessor or
AnnotationAsyncExecutionAspect, respectively.  Because @Configuration
classes cannot conditionally register @Bean methods, this means that
two distinct @Configuration classes must exist - one for proxy-based
advice, one for AspectJ-based advice.

These two types now exist with ProxyAsyncConfiguration and
AspectJAsyncConfiguration.

The next challenge is deciding which @Configuration class should
be imported based user input through the @EnableAsync#mode attribute.

This is done with AsyncConfigurationSelector, an implementation of the
recently introduced ImportSelector interface. AsyncConfigurationSelector
analyzes the value of the #mode attribute and returns the correct
@Configuration to be imported.

Note that AspectJAsyncConfiguration resides in the .aspects module due
to its compile-time dependency on AnnotationAsyncExecutionAspect.
This class must of course use the @Configuration and @Bean annotations,
therefore an explicit dependency from .aspects -> .context has been
created, and template.mf and ivy.xml have been updated to reflect.  This
does not actually represent a serious change, as .aspects has long had
a transitive dependency on .context through the .transaction module.

Also with this change, the EnableAsync#executorName attribute has been
removed in favor of implementing the new 'listener-style'
AsyncConfigurationCustomizer.  See the JavaDoc for full details on how
to use this interface.