SPR-8954 - Changed test case to show lookup by predicted type fails.
When we predict the bean type of a FactoryBean the BeanFactory doesn't return the object created by the factory when accessing it by the type predicted. This is probably due to the lack of a check of a predicted bean type in FactoryBeanRegistrySupport.getTypeForFactoryBean(…) which simply calls factory.getObjectType(). One might assume that this method should return a valid type as the factory bean has been initialized completely at this point but wouldn't it make sense to re-predict in case the factory doesn't return anything for that call?